LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/select-option-focusable.tentative-expected.txt


  one
  two
  three

FAIL Validate <option> is focusable when is a descendant of <select> assert_equals: expected Element node <option id="select0-option2">two</option> but got Element node <select id="select0">
  <option>one</option>
  <option id...
